+++ src/Dao/GenericDAO.java
@@ -9,6 +9,7 @@ import java.util.List;
 /**
  *
  * @author Daniela Nahir Romero
+ * @param <T>
  */
 public interface GenericDAO<T> {
     
@@ -16,9 +17,9 @@ public interface GenericDAO<T> {
     
     void actualizar(T entidad) throws Exception;
     
-    void eliminar(int id) throws Exception;
+    void eliminar(long id) throws Exception;
     
-    T getById(int id) throws Exception;
+    T getById(long id) throws Exception;
     
     List<T> getAll() throws Exception;
