+++ src/Service/GenericService.java
@@ -13,8 +13,8 @@ public interface GenericService<T> {
     //interfaz generica que define las operaciones basicas que deben implementar todos los servicios
     void insertar (T entidad) throws Exception;
     void actualizar (T entidad) throws Exception;
-    void eliminar (int id_dom)throws Exception;
-    T getByID (int id) throws Exception;
+    void eliminar (long id_dom)throws Exception;
+    T getByID (long id) throws Exception;
     List<T> getAll() throws Exception;
          
 }
